+++ spinta/datasets/backends/notimpl/components.py
@@ -1,0 +1,5 @@
+from spinta.datasets.components import ExternalBackend
+
+
+class BackendNotImplemented(ExternalBackend):
+    pass