+++ spinta/dimensions/param/components.py
@@ -1,0 +1,11 @@
+from typing import Any
+from typing import Dict
+
+from spinta.core.ufuncs import Env
+
+
+class ParamBuilder(Env):
+    pass
+
+
+ResolvedParams = Dict[str, Any]